Relax pydantic requirements constraint
Since gluon-ts may be used in code which also depends on pydantic, it
would be nice if it played nice with others. Given that pydantic is
v1.1.0 at the time of this commit, we relax the constraint so that it
will still install as before, but allow up to versions <2 to be
installed as well.