Only run tests in the test directory
Pytest was looking for tests in any file in the project's directory
which would then lead it to find TestSplitSampler in the transform.py
script. However, as this is not a test, we don't want pytest to
collect it. By specifying we only want to run tests from the test
directory, we avoid this issue.